@@ -1,7 +1,7 @@
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.10'
+gem 'rails', '5.1.4'
 ruby '2.4.2'
 
 # Use sqlite3 as the database for Active Record
@@ -35,12 +35,10 @@ gem 'capistrano3-puma'
 
 gem 'capistrano-sidekiq', github: 'seuros/capistrano-sidekiq'
 
-gem 'unobtrusive_flash'
-
 gem 'mailboxer'
 
 gem 'sidekiq'
-gem 'sinatra'
+
 gem 'sshkit-sudo'
 
 group :development do
@@ -61,18 +59,18 @@ group :test do
   gem 'guard-minitest'
 end
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 4.0.3'
+
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # Use CoffeeScript for .js.coffee assets and views
-gem 'coffee-rails', '~> 4.0.0'
+
 # See https://github.com/sstephenson/execjs#readme for more supported runtimes
 # gem 'therubyracer',  platforms: :ruby
 
 # Use jquery as the JavaScript library
-gem 'jquery-rails'
+
 # Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem 'jquery-turbolinks'
+
 gem 'turbolinks'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
 gem 'jbuilder', '~> 2.0'
@@ -109,9 +107,16 @@ gem 'scalpel'
 
 gem 'pg_search'
 
+gem 'sinatra'
 gem 'owlcarousel-rails'
-
+gem 'jquery-turbolinks'
 gem 'public_activity'
+gem 'unobtrusive_flash'
+gem 'sass-rails'
+gem 'jquery-rails'
+gem 'coffee-rails'
+
+gem 'record_tag_helper'
 
 # gem 'omniauth-twitter'
 # Use Capistrano for deployment
