@@ -1,5 +1,5 @@
 class MicropostsController < ApplicationController
-  before_filter :authenticate_user!, only: [:create, :new, :new_user_post, :destroy, :update, :edit]
+  before_action :authenticate_user!, only: [:create, :new, :new_user_post, :destroy, :update, :edit]
   before_action :correct_user, only: :destroy
   before_action :set_micropost, only: [:edit, :new, :update, :destroy, :show]
   before_action :follow, only: :show
